Remove automatically URI-decoding params
Here's why:

Most params don't need to be URI-decoded. Path params are most often an ID made up of only alphanumeric characters. Other common cases are file paths (from wildcards) and human-friendly identifiers that are restricted to characters that don't need to be URI-decoded.

For the cases where a param does need to be decoded, letting the user decode it themselves gives them the freedom to choose how to decode it as well as full control over handling decoding errors——e.g. `decodeURIComponent('throws%URIError')`.